move QueryParamsGetEntitlements
